+++ nix/nixos/cardano-node-service.nix
@@ -24,7 +24,7 @@ let
               TargetNumberOfKnownPeers = cfg.targetNumberOfKnownPeers;
               TargetNumberOfEstablishedPeers = cfg.targetNumberOfEstablishedPeers;
               TargetNumberOfActivePeers = cfg.targetNumberOfActivePeers;
-              TestEnableDevelopmentNetworkProtocols = true;
+              ExperimentalProtocolsEnabled = true;
               MaxConcurrencyBulkSync = 2;
             })) cfg.extraNodeConfig;
         baseInstanceConfig =