+++ nix/workbench/service/nodes.nix
@@ -166,10 +166,10 @@ let
           ]
         //
         {
-          TestEnableDevelopmentHardForkEras     = true;
-          TestEnableDevelopmentNetworkProtocols = true;
-          TurnOnLogMetrics                      = true;
-          SnapshotFrequency                     = 1100;
+          ExperimentalHardForksEnabled = true;
+          ExperimentalProtocolsEnabled = true;
+          TurnOnLogMetrics             = true;
+          SnapshotFrequency            = 1100;
         };
       tracing-transform = {
         trace-dispatcher = cfg: