+++ scripts/byron-to-alonzo/mkfiles.sh
@@ -663,8 +663,8 @@ if [ "$1" = "alonzo" ]; then
   echo "TestAllegraHardForkAtEpoch: 0" >> ${ROOT}/configuration.yaml
   echo "TestMaryHardForkAtEpoch: 0" >> ${ROOT}/configuration.yaml
   echo "TestAlonzoHardForkAtEpoch: 0" >> ${ROOT}/configuration.yaml
-  echo "TestEnableDevelopmentHardForkEras: True" >> ${ROOT}/configuration.yaml
-  echo "TestEnableDevelopmentNetworkProtocols: True" >> ${ROOT}/configuration.yaml
+  echo "ExperimentalHardForksEnabled: True" >> ${ROOT}/configuration.yaml
+  echo "ExperimentalProtocolsEnabled: True" >> ${ROOT}/configuration.yaml
 
   $SED -i ${ROOT}/configuration.yaml \
       -e 's/LastKnownBlockVersion-Major: 1/LastKnownBlockVersion-Major: 5/'